requirements.txt
# ============================================================
# PulseNet — Production Dependencies
# ============================================================

# Core
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0

# ML
scikit-learn>=1.3.0
joblib>=1.3.0
torch>=2.0.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# Security
cryptography>=41.0.0

# Dashboard
streamlit>=1.28.0
plotly>=5.15.0

# MLOps
mlflow>=2.5.0

# Benchmarking & Profiling
psutil>=5.9.0
matplotlib>=3.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0